% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPT.R
\name{ASPT}
\alias{ASPT}
\title{ASPT Function}
\usage{
ASPT(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ?cleanTax() for help.}
}
\description{
Returns ASPT index values to a dataClean dataframe
}
\author{
Shuntaro Koizumi, Christian Bodin, Gaute Velle
}
