% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTax.R
\name{cleanTax}
\alias{cleanTax}
\title{cleanTax Function}
\usage{
cleanTax(datain)
}
\arguments{
\item{datain}{A dataframe containing columns River, Date (Date Class, set as YMD), Species (taxa name), and Value (Abundance)}
}
\value{
dataClean is a dataframe with corrected species names

ErrorList is a list of removed names from the original dataframe
}
\description{
Cleans dataframe based on a list of tax. For the list of taxa use data(Taxonomy).
}
\author{
Shuntaro Koizumi, Christian Bodin, Gaute Velle
}
