% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1f2.R
\name{f1f2}
\alias{f1f2}
\title{F1F2 Function}
\usage{
f1f2(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ?cleanTax() for help.}
}
\description{
Returns F1F2 index values to a dataClean dataframe
}
\author{
Shuntaro Koizumi, Christian Bodin, Gaute Velle
}
