% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lami.R
\name{lami}
\alias{lami}
\title{Lake Acidification Macroinvertebrate Index (LAMI)}
\usage{
lami(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ? cleanTax() for help.}
}
\description{
This package provides functions to calculate the Lake Acidification Macroinvertebrate Index (LAMI) from macroinvertebrate data.
}
\author{
Shuntaro Koizumi, Gaute Velle
}
