% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{taxaCount Function}
\usage{
psi(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ?cleanTax() for help.}
}
\description{
Counts the number of Ephemeroptera, Plecoptera, Trichoptera, and all taxa
}
\author{
Shuntaro Koizumi, Christian Bodin, Gaute Velle
}
