% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rami.R
\name{rami}
\alias{rami}
\title{Rami Function}
\usage{
rami(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ?cleanTax() for help.}
}
\description{
Returns Rami index values to a dataClean dataframe
}
\author{
Shuntaro Koizumi, Christian Bodin, Gaute Velle
}
