% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdi.R
\name{sdi}
\alias{sdi}
\title{Shannon Diversity Index (H) Function}
\usage{
sdi(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ?cleanTax() for help.}
}
\description{
returns the Shannon diversity index
}
\author{
Shuntaro Koizumi, Gaute Velle
}
