% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaCount.R
\name{taxaCount}
\alias{taxaCount}
\title{taxaCount Function}
\usage{
taxaCount(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ?cleanTax() for help.}
}
\description{
Counts the number of Ephemeroptera, Plecoptera, Trichoptera, and all taxa
}
\author{
Shuntaro Koizumi, Christian Bodin, Gaute Velle
}
