% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiClear.R
\name{multiClear}
\alias{multiClear}
\title{Multimetric Invertebrate Index for Clear Lakes (MultiClear)}
\usage{
multiClear(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ? cleanTax() for help.}
}
\description{
Function outputs MultiClear, which is a multimetric index (MMI) for assessing the ecological status of clear lakes in Norway (Sandin et al. 2014).

MultiClear is comprised of four indices. These four indices are:
\enumerate{
\item Total Number of Taxa of Gastropoda
\item Total Number of Taxa of Ephemeroptera
\item Acid Water Indicator Community (AWIC) Index Scores
\item NGIG-Adjusted Henriksson and Medins Index}

The four indices are then combined into a single MMI using the following formula:

\eqn{ \text{MultiClear} =  \frac{(\sum_n^k=1  M_{k}) }{n}}   

where \eqn{M_{k}} is the kth metric, and n is the number of metrics.
}
\author{
Shuntaro Koizumi, Gaute Velle
}
