% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMI.R
\name{nami}
\alias{nami}
\title{NAMI (Nordic Acidity Macroinvertebrate Index)}
\usage{
nami(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ?cleanTax() for help.}
}
\description{
Returns NAMI values based on a set of indices published in Carlson et al. 2023 (Ecological Indicators).

NAMI is an MMI (multimetric index) comprised of seven indices. These seven indices are:
\enumerate{
\item The cumulative proportions of Gastropoda, Bivalvia, and Crustacea taxa 
\item The number of Bivalvia taxa 
\item Number of Ephemeroptera (excluding Leptophlebiidae)
\item Life Cycle Duration (>1 year)
\item Resistance Forms (Egg Stages)
\item Respiration Method (Plastron)
\item pH Preferendum (>5-5.5)
}
 After these seven metrics are calculated, each individual metric is standardized based on the inputted dataset,
 using the following transformation:
 
 \eqn{\text{Value} =  \dfrac{(\text{Metric Result} - \text{Lowest Metric Value})}{(\text{Highest Metric Value} - \text{Lowest Metric Value})}}
 
 These standardized metrics are then combined as the mean of the 0 to 1 scores of all core metrics. 
 
 MMI values are then recalibrated using the 90th and 10th percentiles as upper and lower limits of the metrics calculated 
 within the inputted dataset. Be aware that this assumes that the inputted dataset in @param dataClean has sufficient data
 and contains reference data. 
 
 Note: For metric 1 and 2 in the list above, species name from dataClean are matched with an internal dataframe which contains
 a list of species and associated higher classifications.
}
\author{
Shuntaro Koizumi, Gaute Velle
}
