% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Proportion of Sediment-sensitive Invertebrates (PSI)}
\usage{
psi(dataClean)
}
\arguments{
\item{dataClean}{A dataframe output from cleanTax(). See ?cleanTax() for help.}
}
\description{
Index designed to identify the degree of sedimentation in rivers and streams.
}
\author{
Shuntaro Koizumi, Christian Bodin, Gaute Velle
}
